Feature/#21 포스트 등록, 조회, 삭제, 수정, 전체 포스트 조회

+++ src/main/java/com/ll/netmong/domain/post/service/PostServiceImpl.java
@@ -1,5 +1,7 @@
 package com.ll.netmong.domain.post.service;
 
+import com.ll.netmong.common.PermissionDeniedException;
+import com.ll.netmong.domain.member.entity.Member;
 import com.ll.netmong.domain.post.dto.request.PostRequest;
 import com.ll.netmong.domain.post.dto.response.PostResponse;
 import com.ll.netmong.domain.post.entity.Post;
@@ -8,7 +10,9 @@ import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -17,10 +21,21 @@ public class PostServiceImpl implements PostService {
     private final PostRepository postRepository;
 
     @Override
+    public List<PostResponse> getViewAll() {
+        List<Post> posts = postRepository.findAll();
+
+        return posts.stream()
+                .map(PostResponse::new)
+                .collect(Collectors.toList());
+    }
+
+    @Override
     @Transactional
-    public void uploadPost(MultipartFile image, PostRequest postRequest) {
+    public void uploadPost(PostRequest postRequest, Member foundMember, String foundUsername) {
         postRepository.save(Post.builder()
                 .title(postRequest.getTitle())
+                .member(foundMember)
+                .writer(foundUsername)
                 .content(postRequest.getContent())
                 .imageUrl(postRequest.getImageUrl()).build());
     }
@@ -35,21 +50,32 @@ public class PostServiceImpl implements PostService {
 
     @Override
     @Transactional
-    public void deletePost(long id) {
-        postRepository.deleteById(id);
+    public void deletePost(Long id, String foundUsername) {
+        Post post = postRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("포스트를 찾을 수 없습니다."));
+
+        if (post.getWriter().equals(foundUsername)) {
+            postRepository.deleteById(id);
+        } else {
+            throw new PermissionDeniedException("해당 포스트에 대한 삭제 권한이 없습니다.");
+        }
     }
 
     @Override
     @Transactional
-    public void updatePost(long id, PostRequest updatedPostRequest) {
+    public void updatePost(Long id, PostRequest updatedPostRequest, String foundUsername) {
         Post originPost = postRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("포스트를 찾을 수 없습니다."));
 
-        Post updatedPost = originPost.toBuilder()
-                .title(updatedPostRequest.getTitle())
-                .content(updatedPostRequest.getContent())
-                .build();
+        if (originPost.getWriter().equals(foundUsername)) {
+            Post updatedPost = originPost.toBuilder()
+                    .title(updatedPostRequest.getTitle())
+                    .content(updatedPostRequest.getContent())
+                    .imageUrl(updatedPostRequest.getImageUrl()).build();
 
-        postRepository.save(updatedPost);
+            postRepository.save(updatedPost);
+        } else {
+            throw new PermissionDeniedException("해당 포스트에 대한 수정 권한이 없습니다.");
+        }
     }
 }
